+++ pyxem/signals/indexation_results.py
@@ -75,25 +75,6 @@ def crystal_from_vector_matching(z_matches):
 
     return results_array
 
-
-def get_phase_name_and_index(library):
-    """Get a dictionary of phase names and its corresponding index value in library.keys().
-
-    Parameters
-    ----------
-    library : DiffractionLibrary
-        Diffraction library containing the phases and rotations
-
-    Returns
-    -------
-    phase_name_index_dict : Dictionary {str : int}
-    typically on the form {'phase_name 1' : 0, 'phase_name 2': 1, ...}
-    """
-
-    phase_name_index_dict = dict([(y, x) for x, y in enumerate(list(library.keys()))])
-    return phase_name_index_dict
-
-
 def _peaks_from_best_template(single_match_result, library, rank=0):
     """Takes a TemplateMatchingResults object and return the associated peaks,
     to be used in combination with map().
@@ -117,7 +98,7 @@ def _peaks_from_best_template(single_match_result, library, rank=0):
     phase_names = list(library.keys())
     phase_index = int(best_fit[0])
     phase = phase_names[phase_index]
-    simulation = library.get_library_entry(phase=phase, angle=tuple(best_fit[1]))["Sim"]
+    simulation = library.get_library_entry(phase=phase, angle=tuple(best_fit[1:4]))["Sim"]
 
     peaks = simulation.coordinates[:, :2]  # cut z
     return peaks
@@ -213,7 +194,7 @@ class TemplateMatchingResults(GenericMatchingResults):
         **kwargs :
             Keyword arguments passed to signal.plot()
         """
-        match_peaks = self.map(_peaks_from_best_template, library=library, inplace=False)
+        match_peaks = self.data.map(_peaks_from_best_template, library=library, inplace=False)
         mmx, mmy = generate_marker_inputs_from_peaks(match_peaks)
         signal.plot(*args, **kwargs)
         for mx, my in zip(mmx, mmy):